# Conflicts:
#	app/src/main/java/com/example/sdo_project/MainActivity.kt
#	app/src/main/java/com/example/sdo_project/data/remote/dto/DisciplineDto.kt
#	app/src/main/java/com/example/sdo_project/data/repository/SupabaseDisciplineRepositoryImpl.kt
#	app/src/main/java/com/example/sdo_project/data/repository/UserRepositoryImpl.kt
#	app/src/main/java/com/example/sdo_project/di/AppModule.kt
#	app/src/main/java/com/example/sdo_project/domain/models/Discipline.kt
#	app/src/main/java/com/example/sdo_project/domain/usecase/auth/AuthUseCases.kt
#	app/src/main/java/com/example/sdo_project/domain/usecase/grade/GetSectionGradesByStudentId.kt
#	app/src/main/java/com/example/sdo_project/domain/usecase/grade/GetSectionsGradesByStudentId.kt
#	app/src/main/java/com/example/sdo_project/domain/usecase/user/EditUserUseCase.kt
#	app/src/main/java/com/example/sdo_project/domain/usecase/user/GetUserUseCase.kt
#	app/src/main/java/com/example/sdo_project/presentation/MainEvent.kt
#	app/src/main/java/com/example/sdo_project/presentation/MainState.kt
#	app/src/main/java/com/example/sdo_project/presentation/auth/AuthScreen.kt
#	app/src/main/java/com/example/sdo_project/presentation/home_navigator/HomeNavigatorScreen.kt
#	app/src/main/java/com/example/sdo_project/presentation/navgraph/NavGraph.kt
#	local.properties

+++ app/src/main/java/com/example/sdo_project/data/remote/dto/GradePointDto.kt
@@ -8,6 +8,6 @@ import kotlinx.serialization.Serializable
 data class GradePointDto(
     @SerialName("point_id") val id: Int,
     @SerialName("point_name") val name: String,
-    @SerialName("point_current_score") val currentScore: Int,
+    @SerialName("point_current_score") val currentScore: Float,
     @SerialName("point_max_score") val maxScore: Int
 )
